+++ tests/test_lib/polyfill/test_h3.py
@@ -119,63 +119,63 @@ def test_polygon_to_cells():
 
 def test_polygon_to_cells_experimental():
     poly = h3.LatLngPoly(sf_7x7)
-    for flags in [0, 'containment_center', h3.ContainmentMode.containment_center]:
-        # Note that `polygon_to_cells` is an alias for `h3shape_to_cells`
-        out = h3.polygon_to_cells_experimental(poly, res=9, flags=flags)
+    # Note that `polygon_to_cells_experimental` is an alias for
+    # `h3shape_to_cells_experimental`
+    out = h3.polygon_to_cells_experimental(poly, res=9, contain='center')
 
-        assert len(out) == 1253
-        assert '89283080527ffff' in out
-        assert '89283095edbffff' in out
+    assert len(out) == 1253
+    assert '89283080527ffff' in out
+    assert '89283095edbffff' in out
 
 
-def test_polygon_to_cells_experimental_full():
+def test_h3shape_to_cells_experimental():
     poly = h3.LatLngPoly(sf_7x7)
-    for flags in [1, 'containment_full', h3.ContainmentMode.containment_full]:
-        # Note that `polygon_to_cells` is an alias for `h3shape_to_cells`
-        out = h3.polygon_to_cells_experimental(poly, res=9, flags=flags)
 
-        assert len(out) == 1175
-        assert '89283082a1bffff' in out
-        assert '89283080527ffff' not in out
-        assert '89283095edbffff' in out
+    out = h3.h3shape_to_cells_experimental(poly, res=9, contain='center')
+
+    assert len(out) == 1253
+    assert '89283080527ffff' in out
+    assert '89283095edbffff' in out
 
 
-def test_polygon_to_cells_experimental_overlapping():
+def test_h3shape_to_cells_experimental_full():
     poly = h3.LatLngPoly(sf_7x7)
-    for flags in [
-        2,
-        'containment_overlapping',
-        h3.ContainmentMode.containment_overlapping
-    ]:
-        # Note that `polygon_to_cells` is an alias for `h3shape_to_cells`
-        out = h3.polygon_to_cells_experimental(poly, res=9, flags=flags)
 
-        assert len(out) == 1334
-        assert '89283080527ffff' in out
-        assert '89283095edbffff' in out
+    out = h3.h3shape_to_cells_experimental(poly, res=9, contain='full')
+
+    assert len(out) == 1175
+    assert '89283082a1bffff' in out
+    assert '89283080527ffff' not in out
+    assert '89283095edbffff' in out
 
 
-def test_polygon_to_cells_experimental_overlapping_bbox():
+def test_h3shape_to_cells_experimental_overlapping():
     poly = h3.LatLngPoly(sf_7x7)
-    for flags in [
-        3,
-        'containment_overlapping_bbox',
-        h3.ContainmentMode.containment_overlapping_bbox
-    ]:
-        # Note that `polygon_to_cells` is an alias for `h3shape_to_cells`
-        out = h3.polygon_to_cells_experimental(poly, res=9, flags=flags)
 
-        assert len(out) == 1416
-        assert '89283080527ffff' in out
-        assert '89283095edbffff' in out
+    out = h3.h3shape_to_cells_experimental(poly, res=9, contain='overlap')
+
+    assert len(out) == 1334
+    assert '89283080527ffff' in out
+    assert '89283095edbffff' in out
 
 
-def test_polygon_to_cells_experimental_invalid_mode():
+def test_h3shape_to_cells_experimental_overlapping_bbox():
     poly = h3.LatLngPoly(sf_7x7)
-    for flags in [1.0, 'containment_overlapping_bbox_abc', None]:
-        with pytest.raises(ValueError):
-            # Note that `polygon_to_cells` is an alias for `h3shape_to_cells`
-            h3.polygon_to_cells_experimental(poly, res=9, flags=flags)
+    out = h3.h3shape_to_cells_experimental(poly, res=9, contain='bbox_overlap')
+
+    assert len(out) == 1416
+    assert '89283080527ffff' in out
+    assert '89283095edbffff' in out
+
+
+def test_h3shape_to_cells_experimental_invalid_mode():
+    poly = h3.LatLngPoly(sf_7x7)
+    with pytest.raises(KeyError):
+        h3.h3shape_to_cells_experimental(
+            poly,
+            res = 9,
+            contain = 'contain_overlapping_bbox_abc',
+        )
 
 
 def test_poly_to_cells_experimental_mpoly():
@@ -185,25 +185,25 @@ def test_poly_to_cells_experimental_mpoly():
     )
 
     assert (
-        set(h3.polygon_to_cells_experimental(mpoly, res=9))
+        set(h3.h3shape_to_cells_experimental(mpoly, res=9))
         ==
-        set(h3.polygon_to_cells_experimental(mpoly, res=9, flags='containment_center'))
+        set(h3.h3shape_to_cells_experimental(mpoly, res=9, contain='center'))
     )
 
     assert (
-        set(h3.polygon_to_cells_experimental(mpoly, res=9))
+        set(h3.h3shape_to_cells_experimental(mpoly, res=9))
         <
-        set(h3.polygon_to_cells_experimental(
+        set(h3.h3shape_to_cells_experimental(
             mpoly,
-            res=9,
-            flags='containment_overlapping'
+            res = 9,
+            contain = 'overlap'
         ))
     )
 
-    assert 120 == len(h3.polygon_to_cells_experimental(
+    assert 120 == len(h3.h3shape_to_cells_experimental(
         mpoly,
-        res=9,
-        flags='containment_overlapping'
+        res = 9,
+        contain = 'overlap'
     ))